Use efficient page-writes for power loss recovery writes to EEPROM
Writing power-loss recovery data was taking about 0.5s, which was enough in some conditions (particularly vase mode) to cause the head movement to unintentionally pause causing blobbing.
This was caused by doing single-byte writes with a 10ms write cycle between each byte of the 184-byte recovery structure to EEPROM.
This is a rewrite of the block write routine to use more efficient page writes - takes a little over 60ms to do the same according to my calculations (the time taken is mostly a 5ms write cycle every 16 bytes).