Remove the behaviour that always homes XY at the end of a print
The Creality firmware hard-codes the system to home XY and disable steppers at the end of an SD card print. I prefer to put such things in the slicer's printer end g-code, where I can control what happens at the end of a print - in particular to move the bed fowards rather than back to make it easy to pick up the finished print, however that is overridden by this behaviour..
This change adds a configuration macro to easily remove this behaviour.