Comment correct datasheet for the EEPROM device used on the Creality 4.5.2 board
I used the BL24C16 generic datasheet to develop the page write code, however the precise device is a BL24C16A in a TSOT23-5 package. This is described in a different datasheet - luckily the protocol is the same so no code change necessary, but useful to reference the correct datasheet.